Move used regs to utils
