+++ utils.c
@@ -1,8 +1,10 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <stdbool.h>
+#include "var.h"
 
-// Remove all occurances of c from str
+// Remove all occurrences of c from str
 void removeChars(char* str, char c) {
     char *pr = str, *pw = str;
     while (*pr) {
@@ -24,4 +26,52 @@ int getOperands(char** operandList, char* operandString) {
         operand = strtok(NULL, ",");
     }
     return numOperands;
+}
+
+bool exists(char* reg, char** usedRegs, int count){
+    for (int i = 0 ; i < count ; i++){
+        if(strcmp(usedRegs[i],reg) == 0){
+            return true;
+        }
+    }
+    return false;
+}
+
+// return a list of registers currently in use
+char** usedRegisters(Vars* vars){
+    char** usedRegs = malloc(vars->count * sizeof(char*));
+    for (int i = 0 ; i < vars->count ; i++) {
+        usedRegs[i] = malloc(4);
+        strcpy(usedRegs[i], vars->vars[i]->reg);
+    }
+    return usedRegs;
+}
+
+void freeUsedRegs(char** usedRegs, int count){
+    for (int i = 0 ; i < count ; i++) {
+        free(usedRegs[i]);    
+    }
+    free(usedRegs);
+}
+
+void addRegToUsed(char** used, char* reg, int count){
+    if(!exists(reg,used,count)){
+        for (int i = 0 ; i < count ; i++) {
+            if(strcmp(used[i],"new") == 0) {
+                strcpy(used[i],reg);
+                return;
+            }
+        }
+    }
+}
+
+void removeRegFromUsed(char** used, char* reg, int count){
+    if(exists(reg,used,count)){
+        for (int i = 0 ; i < count ; i++) {
+            if(strcmp(used[i],reg) == 0) {
+                strcpy(used[i],"new");
+                return;
+            }
+        }
+    }
 }
